+++ src/features/home/components/sections/InvestmentSection.tsx
