+++ src/features/home/components/sections/InvestorSection.tsx
