+++ src/features/home/hooks/useHomeStaticQuery.ts
